refactor[#232] : PetInfoForm mutation option 이동
